requirements.txt
# Core SNAFU dependencies
numpy>=1.21.0
networkx>=3.0
more_itertools
planarity

# Data science and analysis
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Natural language processing
spacy>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Statistical analysis
statsmodels>=0.13.0

# Configuration and utilities
pyyaml>=6.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.0.0
black>=21.0.0
flake8>=3.8.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Optional: For advanced statistical analysis
# pingouin>=0.5.0  # Uncomment if you want additional statistical tests
